+++ providers/dns/transip/transip.go
@@ -5,12 +5,10 @@ import (
 	"errors"
 	"fmt"
 	"strings"
-	"sync"
 	"time"
 
 	"github.com/go-acme/lego/v3/challenge/dns01"
 	"github.com/go-acme/lego/v3/platform/config/env"
-
 	"github.com/transip/gotransip/v6"
 	transipdomain "github.com/transip/gotransip/v6/domain"
 )
@@ -47,9 +45,8 @@ func NewDefaultConfig() *Config {
 
 // DNSProvider describes a provider for TransIP
 type DNSProvider struct {
-	config       *Config
-	repository   transipdomain.Repository
-	dnsEntriesMu sync.Mutex
+	config     *Config
+	repository transipdomain.Repository
 }
 
 // NewDNSProvider returns a DNSProvider instance configured for TransIP.
@@ -107,10 +104,6 @@ func (d *DNSProvider) Present(domain, token, keyAuth string) error {
 	// get the subDomain
 	subDomain := strings.TrimSuffix(dns01.UnFqdn(fqdn), "."+domainName)
 
-	// use mutex to prevent race condition from GetInfo until SetDNSEntries
-	d.dnsEntriesMu.Lock()
-	defer d.dnsEntriesMu.Unlock()
-
 	entry := transipdomain.DNSEntry{
 		Name:    subDomain,
 		Expire:  int(d.config.TTL),
@@ -119,10 +112,10 @@ func (d *DNSProvider) Present(domain, token, keyAuth string) error {
 	}
 
 	err = d.repository.AddDNSEntry(domainName, entry)
-
 	if err != nil {
 		return fmt.Errorf("transip: %w", err)
 	}
+
 	return nil
 }
 
@@ -139,10 +132,6 @@ func (d *DNSProvider) CleanUp(domain, token, keyAuth string) error {
 
 	// get the subDomain
 	subDomain := strings.TrimSuffix(dns01.UnFqdn(fqdn), "."+domainName)
-
-	// use mutex to prevent race condition from GetInfo until SetDNSEntries
-	d.dnsEntriesMu.Lock()
-	defer d.dnsEntriesMu.Unlock()
 
 	// get all DNS entries
 	dnsEntries, err := d.repository.GetDNSEntries(domainName)
