+++ providers/dns/transip/transip_test.go
@@ -1,7 +1,6 @@
 package transip
 
 import (
-	"encoding/json"
 	"errors"
 	"fmt"
 	"os"
@@ -10,129 +9,11 @@ import (
 	"testing"
 	"time"
 
-	"github.com/transip/gotransip/v6/rest"
-
 	"github.com/go-acme/lego/v3/platform/tester"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"github.com/transip/gotransip/v6/domain"
 )
-
-type fakeClient struct {
-	dnsEntries           []domain.DNSEntry
-	setDNSEntriesLatency time.Duration
-	getInfoLatency       time.Duration
-	domainName           string
-}
-
-type dnsEntryWrapper struct {
-	DNSEntry domain.DNSEntry `json:"dnsEntry"`
-}
-
-type dnsEntriesWrapper struct {
-	DNSEntries []domain.DNSEntry `json:"dnsEntries"`
-}
-
-func (f *fakeClient) Get(request rest.Request, dest interface{}) error {
-	if f.getInfoLatency != 0 {
-		time.Sleep(f.getInfoLatency)
-	}
-
-	switch request.Endpoint {
-	case fmt.Sprintf("/domains/%s/dns", f.domainName):
-		entries := dnsEntriesWrapper{DNSEntries: f.dnsEntries}
-		body, err := json.Marshal(entries)
-
-		if err != nil {
-			return fmt.Errorf("can't encode json: %w", err)
-		}
-
-		err = json.Unmarshal(body, dest)
-
-		if err != nil {
-			return fmt.Errorf("can't decode json: %w", err)
-		}
-	default:
-		return fmt.Errorf("function GET for endpoint %s not implemented", request.Endpoint)
-	}
-
-	return nil
-}
-
-func (f *fakeClient) Put(request rest.Request) error {
-	if f.getInfoLatency != 0 {
-		time.Sleep(f.getInfoLatency)
-	}
-
-	return fmt.Errorf("function PUT for endpoint %s not implemented", request.Endpoint)
-}
-
-func (f *fakeClient) Post(request rest.Request) error {
-	if f.getInfoLatency != 0 {
-		time.Sleep(f.getInfoLatency)
-	}
-	switch request.Endpoint {
-	case fmt.Sprintf("/domains/%s/dns", f.domainName):
-		body, err := request.GetJSONBody()
-		if err != nil {
-			return fmt.Errorf("unable get request body")
-		}
-
-		var entry dnsEntryWrapper
-		if err := json.Unmarshal(body, &entry); err != nil {
-			return fmt.Errorf("unable to decode request body")
-		}
-
-		f.dnsEntries = append(f.dnsEntries, entry.DNSEntry)
-	default:
-		return fmt.Errorf("function POST for endpoint %s not implemented", request.Endpoint)
-	}
-
-	return nil
-}
-
-func (f *fakeClient) Delete(request rest.Request) error {
-	if f.getInfoLatency != 0 {
-		time.Sleep(f.getInfoLatency)
-	}
-
-	switch request.Endpoint {
-	case fmt.Sprintf("/domains/%s/dns", f.domainName):
-		body, err := request.GetJSONBody()
-		if err != nil {
-			return fmt.Errorf("unable get request body")
-		}
-
-		var entry dnsEntryWrapper
-		if err := json.Unmarshal(body, &entry); err != nil {
-			return fmt.Errorf("unable to decode request body")
-		}
-
-		cp := make([]domain.DNSEntry, 0)
-
-		for _, e := range f.dnsEntries {
-			if e.Name == entry.DNSEntry.Name {
-				continue
-			}
-
-			cp = append(cp, e)
-		}
-
-		f.dnsEntries = cp
-	default:
-		return fmt.Errorf("function DELETE for endpoint %s not implemented", request.Endpoint)
-	}
-
-	return nil
-}
-
-func (f *fakeClient) Patch(request rest.Request) error {
-	if f.getInfoLatency != 0 {
-		time.Sleep(f.getInfoLatency)
-	}
-
-	return fmt.Errorf("function PATCH for endpoint %s not implemented", request.Endpoint)
-}
 
 const envDomain = envNamespace + "DOMAIN"
 
@@ -200,8 +81,8 @@ func TestNewDNSProvider(t *testing.T) {
 		})
 	}
 
-	// The error message for a file not existing is different on Windows and Linux. Therefore we test if the error
-	// type is the same
+	// The error message for a file not existing is different on Windows and Linux.
+	// Therefore we test if the error type is the same.
 	t.Run("could not open private key path", func(t *testing.T) {
 		defer envTest.RestoreEnv()
 		envTest.ClearEnv()
@@ -213,7 +94,7 @@ func TestNewDNSProvider(t *testing.T) {
 
 		_, err := NewDNSProvider()
 		if !errors.Is(err, os.ErrNotExist) {
-			t.Fatalf("Expected an os.ErrNotExists error\nActual: %v", err)
+			t.Fatalf("Expected an os.ErrNotExists error, actual: %v", err)
 		}
 	})
 }
@@ -265,8 +146,8 @@ func TestNewDNSProviderConfig(t *testing.T) {
 		})
 	}
 
-	// The error message for a file not existing is different on Windows and Linux. Therefore we test if the error
-	// type is the same
+	// The error message for a file not existing is different on Windows and Linux.
+	// Therefore we test if the error type is the same.
 	t.Run("could not open private key path", func(t *testing.T) {
 		config := NewDefaultConfig()
 		config.AccountName = "johndoe"
@@ -275,12 +156,12 @@ func TestNewDNSProviderConfig(t *testing.T) {
 		_, err := NewDNSProviderConfig(config)
 
 		if !errors.Is(err, os.ErrNotExist) {
-			t.Fatalf("Expected an os.ErrNotExists error\nActual: %v", err)
+			t.Fatalf("Expected an os.ErrNotExists error, actual: %v", err)
 		}
 	})
 }
 
-func TestDNSProvider_concurrentGetInfo(t *testing.T) {
+func TestDNSProvider_concurrentGetDNSEntries(t *testing.T) {
 	client := &fakeClient{
 		getInfoLatency:       50 * time.Millisecond,
 		setDNSEntriesLatency: 500 * time.Millisecond,
@@ -299,12 +180,14 @@ func TestDNSProvider_concurrentGetInfo(t *testing.T) {
 
 	solve := func(domain1 string, suffix string, timeoutPresent time.Duration, timeoutSolve time.Duration, timeoutCleanup time.Duration) error {
 		time.Sleep(timeoutPresent)
+
 		err := p.Present(domain1, "", "")
 		if err != nil {
 			return err
 		}
 
 		time.Sleep(timeoutSolve)
+
 		var found bool
 		for _, entry := range client.dnsEntries {
 			if strings.HasSuffix(entry.Name, suffix) {
@@ -316,6 +199,7 @@ func TestDNSProvider_concurrentGetInfo(t *testing.T) {
 		}
 
 		time.Sleep(timeoutCleanup)
+
 		return p.CleanUp(domain1, "", "")
 	}
 
@@ -336,7 +220,7 @@ func TestDNSProvider_concurrentGetInfo(t *testing.T) {
 	assert.Empty(t, client.dnsEntries)
 }
 
-func TestDNSProvider_concurrentSetDNSEntries(t *testing.T) {
+func TestDNSProvider_concurrentAddDNSEntry(t *testing.T) {
 	client := &fakeClient{
 		domainName: "lego.wtf",
 	}
